Move static methods and strings into an extensions files
