+++ DfE.FindInformationAcademiesTrusts/Pages/Shared/_Footer.cshtml
@@ -8,7 +8,7 @@
             <a asp-page="/Privacy" class="govuk-footer__link">Privacy</a>
           </li>
           <li class="govuk-footer__inline-list-item">
-            <a asp-page="/Cookies" class="govuk-footer__link" asp-route-returnPath="@(string.IsNullOrWhiteSpace(Context.Request.Query["returnPath"]) ? Context.Request.Path + Context.Request.QueryString : Context.Request.Query["returnPath"])">Cookies</a>
+            <a asp-page="/Cookies" class="govuk-footer__link" asp-route-returnPath="@CookiesExtensions.ReturnPath(Context)">Cookies</a>
           </li>
         </ul>
         <svg class="govuk-footer__licence-logo" xmlns="http://www.w3.org/2000/svg" viewbox="0 0 483.2 195.7" height="17" width="41">
