+++ DfE.FindInformationAcademiesTrusts/Pages/_ViewImports.cshtml
@@ -1,5 +1,6 @@
 ﻿@using DfE.FindInformationAcademiesTrusts
 @using DfE.FindInformationAcademiesTrusts.Options
+@using DfE.FindInformationAcademiesTrusts.Extensions
 @using Microsoft.ApplicationInsights.AspNetCore
 @using Microsoft.Extensions.Options
 @namespace DfE.FindInformationAcademiesTrusts.Pages
